% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Contour plot.R
\name{contour.plot}
\alias{contour.plot}
\title{Contour plot of a biogeochemical variable.}
\usage{
\method{contour}{plot}(
  sim = temp.interpolated,
  obs = obs.temp,
  file_name = "Contour_temp.png",
  start.date = "2017-06-06",
  end.date = "2020-02-29",
  date = NULL,
  legend.title = "T\\n(°C)",
  min.depth = 0,
  max.depth = 33,
  by.value = 0.5,
  nlevels = 20,
  height = 5,
  width = 8,
  ppi = 150
)
}
\arguments{
\item{sim}{a matrix of simulated variabls that have been interpolated}

\item{obs}{observed values of variable.}

\item{file_name}{the file path to save the generated contour figure.}

\item{start.date, end.date}{the beginning and ending simulation dates for the intended DYRESM-CAEDYM model run. The date format must be "%Y-%m-%d".}

\item{legend.title}{the legend title of the contour figure.}

\item{min.depth, max.depth}{minimum and maximum depth used to be the start of y axis of the contour plot.}

\item{height, width}{the relative height/width of the figure.}

\item{ppi}{the ppi value of the figure.}
}
\value{
a graph file of contour plot saved in the Figure folder.
}
\description{
Contour plot a matrix of a biogeochemical variable values, which can be generated through "interpol" function.
}
