% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Delete space.R
\name{delete.space}
\alias{delete.space}
\title{Delete all whitespace until a non-whitespace character.}
\usage{
delete.space(extract_val)
}
\arguments{
\item{extract_val}{a vector.}
}
\description{
Delete all whitespace until a non-whitespace character.
}
