% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Delete-space.R
\name{delete_space}
\alias{delete_space}
\title{Delete all whitespace until a non-whitespace character.}
\usage{
delete_space(extract_val)
}
\arguments{
\item{extract_val}{a vector.}
}
\description{
Delete all whitespace until a non-whitespace character.
}
