% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Extract simulation.R
\name{extract.output}
\alias{extract.output}
\title{Extract simulations from DYRESM-CAEDYM output.}
\usage{
\method{extract}{output}(
  dycd.output = "DYsim.nc",
  var.extract = c("TEMP", "DO", "TP", "TN", "NO3", "NH4", "CHLA")
)
}
\arguments{
\item{dycd.output}{the output netcdf file of DYRESM-CAEDYM model}

\item{var.extract}{a vector of variables to be extracted from the output file.
Options include TEMP, DO, TP, NO3, NH4, TN, CHLOR, FDIAT, SSOL1, SSOL2, and PO4.
Apart from nominated variables, simulation period and layer height data are also extracted.}
}
\value{
a list of values of those variables of interst,
   as well as two compulsory variables (i.e. simulation period, water level)
}
\description{
Extract simulations from DYRESM-CAEDYM output. It is recommended to use the following example code to
assign values of each simulated variable to a corresponding matrix.
}
\examples{
var.values<-extract.output(dycd.output = "Data/DYRESM-CAEDYM/DYsim.nc",
                           var.extract = c("TEMP","DO","TP","TN","NO3","NH4","PO4","Chla"),
                           output.name = output.name)

for(i in 1:length(var.values)){
  expres<-paste0(names(var.values)[i],"<-data.frame(var.values[[",i,"]])")
  eval(parse(text=expres))
}
}
