% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Height2Depth.R
\name{hgt.to.dpt}
\alias{hgt.to.dpt}
\title{convert from height to depth}
\usage{
hgt.to.dpt(height)
}
\arguments{
\item{height}{a vector of height profile}
}
\description{
convert from height to depth
}
