% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Height2Depth.R
\name{hgt_to_dpt}
\alias{hgt_to_dpt}
\title{convert from height to depth}
\usage{
hgt_to_dpt(height)
}
\arguments{
\item{height}{a vector of height profile}
}
\description{
convert from height to depth
}
