% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_temp.R
\docType{data}
\name{obs_temp}
\alias{obs_temp}
\title{Example observed profiling temperature data across different depths
 over the period of 6-11 June 2017.}
\format{
A data frame with 77 rows and 3 variables:
\describe{
\item{Date}{date when the monitoring happened}
\item{Depth}{depth of monitoring}
\item{TEMP}{temperature value}
}
}
\source{
self-made.
}
\usage{
data(obs_temp)
}
\description{
A table has three columns. The first column name is Date
in the form of dd-mm-YY.
The second column is Depth where the temperature data was monitored.
The third column is monitored temperature value.
}
\keyword{datasets}
