% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_name.R
\docType{data}
\name{output_name}
\alias{output_name}
\title{Default DYCD simulation variable names with their variable name}
\format{
A data frame with 65 rows and 2 variables:
\describe{
\item{var.name}{variable name}
\item{output.name}{default DYCD simulation variable name}
}
}
\source{
self-made.
}
\usage{
data(output_name)
}
\description{
A table has two columns. The first column name is var.name, meaning variable names that are used in the extract.output function.
The second column is the default DYCD simulation variable names, such as "dyresmLAYER_HTS_Var".
}
\keyword{datasets}
