% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Profile-plot.R
\name{profile.plot}
\alias{profile.plot}
\title{Profile plot of simulated variable values vs. depth}
\usage{
\method{profile}{plot}(
  sim = temp.interpolated,
  obs = obs.temp,
  sim.start = "2017-06-06",
  sim.end = "2020-02-29",
  plot.start = "2017-06-06",
  plot.end = "2020-02-29",
  file_name,
  xlabel = "Temperature °C",
  min.depth = 0,
  max.depth,
  by.value,
  height = 11,
  width = 18
)
}
\arguments{
\item{sim}{interpolated values of variable.}

\item{obs}{observed values of variable.}

\item{sim.start, sim.end}{the beginning and ending simulation dates for the intended DYRESM-CAEDYM model run. The date format must be "\%Y-\%m-\%d".}

\item{plot.start, plot.end}{the beginning and ending dates for the plotting purpose. The date format must be "\%Y-\%m-\%d".}

\item{file_name}{the file path to save the generated profile figure}

\item{xlabel}{the x axis label of the profile figure}

\item{min.depth, max.depth, by.value}{minimum and maximum depth for the profile plot at the depth increment of by.value.}

\item{height, width}{the height and width of the profile figure.}
}
\value{
a profile plot of sim~depth
}
\description{
A post-processing funciton used to visualise model output in a profile graph.
}
