% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Scatter plot.R
\name{scatter.plot}
\alias{scatter.plot}
\title{Scatter plot of sim and obs var values}
\usage{
scatter.plot(
  sim = temp.interpolated,
  obs = obs.temp,
  sim.start = "2017-06-06",
  sim.end = "2020-02-29",
  plot.start = "2017-06-06",
  plot.end = "2020-02-29",
  file_name,
  min.depth = 0,
  max.depth,
  by.value,
  height = 4,
  width = 7
)
}
\arguments{
\item{sim}{interpolated values of variable.}

\item{obs}{observed values of variable. This data need to have fixed types of colnames and orders.}

\item{sim.start, sim.end}{the beginning and ending simulation dates for the intended DYRESM-CAEDYM model run. The date format must be "%Y-%m-%d".}

\item{plot.start, plot.end}{the beginning and ending dates for the plotting purpose. The date format must be "%Y-%m-%d".}

\item{file_name}{the file path to save the generated scatter plot.}

\item{height, width}{the height and width of the scatter figure.}
}
\value{
a scatter plot of sim vs. obs
}
\description{
Scatter plot of sim and obs var values
}
