% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Change input file.R
\name{change_input_file}
\alias{change_input_file}
\title{change parameter value of input files to DYRESM_CAEDYM model.}
\usage{
change_input_file(input_file = "par", row_no = 12, new_value)
}
\arguments{
\item{input_file}{vector of input format, such as "par","cfg".}

\item{row_no}{the number of row where the variable of interest is in the input file}

\item{new_value}{the new value that will be assigned to the variable of interest.}
}
\description{
change parameter value of input files to DYRESM_CAEDYM model.
}
