% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_Interpol.R
\name{interpol}
\alias{interpol}
\title{Interpolation of simulation along depths.}
\usage{
interpol(layerHeights, var, min.depth, max.depth, by.value)
}
\arguments{
\item{layerHeights}{layer heights, extracted from DYCD outputs}

\item{var}{simulated variable values, extracted from DYCD outputs}

\item{min.depth, max.depth, by.value}{minimum and maximum depth for interpolation at the depth increment of by.value.}
}
\value{
a matrix of interpolated values of such variable.
}
\description{
Convert simulated variable at irregular layer heights to a dataframe of the same variable at regular (i.e.sequenced) layer heights.
}
