% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun_TS-plot.R
\name{ts.plot}
\alias{ts.plot}
\title{Time series plot of simulated and observed values}
\usage{
ts.plot(
  sim = temp.interpolated,
  obs = obs.temp,
  file_name,
  target.depth = c(1, 6, 12, 30),
  sim.start = "2017-06-06",
  sim.end = "2020-02-29",
  plot.start = "2017-06-06",
  plot.end = "2020-02-29",
  min.depth = 0,
  max.depth = 33,
  by.value = 0.5,
  ylabel = "Temperature °C",
  height = 7,
  width = 11
)
}
\arguments{
\item{sim}{interpolated values of variable}

\item{obs}{observed values of variable}

\item{file_name}{the file path to save the generated ts plot.}

\item{target.depth}{a vector of depth (unit:m) to be used to extract and plot variable values.}

\item{sim.start, sim.end}{the beginning and ending simulation dates for the intended DYRESM-CAEDYM model run. The date format must be "\%Y-\%m-\%d".}

\item{plot.start, plot.end}{the beginning and ending dates for the plotting purpose. The date format must be "\%Y-\%m-\%d".}

\item{min.depth, max.depth, by.value}{minimum and maximum depth for the profile plot at the depth increment of by.value.}

\item{ylabel}{the y axis title.}

\item{height, width}{the height and width of the time series figure.}
}
\value{
a plot of sim and obs time series.


example data can be found in the GitHub repository https://github.com/SongyanYu/dycdTools.
}
\description{
Time series plot of simulated and observed values
}
